+++ python-pack-assign/scripts/data_ingest_run.py
@@ -10,7 +10,7 @@ from python_package.data_ingestion import (
     load_housing_data,
 )
 from sklearn.model_selection import StratifiedShuffleSplit, train_test_split
-from python_package.logging_utils import setup_logging
+from utils.logging_utils import setup_logging
 
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
